adds trees to atlas (#6442)
## About The Pull Request

removes lack of trees to atlas.

obviously can't take screenshots of the whole thing but i will throw
some here


![image](https://github.com/Citadel-Station-13/Citadel-Station-13-RP/assets/54826962/423231d1-42fe-44ee-9401-e457de704ba5)


![image](https://github.com/Citadel-Station-13/Citadel-Station-13-RP/assets/54826962/1afc2500-840c-4cf9-afe0-a16e546dffc6)


![image](https://github.com/Citadel-Station-13/Citadel-Station-13-RP/assets/54826962/448f7ba2-c549-4206-bd89-b34c5800c1c5)


in the western planes, also fixes an invisible door leading down to a
bunker area and fixes the weird white walls in the bomb site


![image](https://github.com/Citadel-Station-13/Citadel-Station-13-RP/assets/54826962/e6bf1ca8-2c4a-40cb-82e2-f8aa4f59b8d5)


## Why It's Good For The Game

greater visual fidelity is good, probably.

## Changelog

:cl:
add: trees to atlas
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->