moving MRE vendors-ATLAS version (WIP) (#6451)
moving the MRE vendor to not be a middle finger to the chef (I was told
I couldn't just delete it, otherwise this would just be a removal.)

<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Removes the MRE vendor in the bar area, shoves it where 2 tables where
on surface 1 in a side room so it's still available for those that use
it.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

I feel like just removing it would be better, but I was told to keep it.
Anyhow for moving it out of the bar is done purely because it being
right next to the kitchen is basically the biggest middle finger you can
give to any chef players. They're already fighting the almighty power
that is the fitmax vendor, adding another reason to not interact with
them 2 feet away from them is just salt in the wound. If you're going to
the bar/kitchen for food items, you are going to interact with service
staff. If you don't wanna interact with service staff, or there are
none, that's what the 50 thousand vendors everywhere else on the map are
for.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Changelog


<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
tweak: Removes MRE vendor in the bar on surface 3
tweak: replaces 2 tables in the side room on surface 1 at the end of the
hall with the MRE vendor
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->