Adds dry monster tamer, some sodas to mining vendor (#6461)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

1. Adds an alternative to monster tamer with the same effects, made with
root beer instead of whiskey. This is almost exactly as easy to obtain;
both are in dispensers, the real "hard part" is grinding up meat.
2. Adds sarsaparilla, root beer and hard root beer to the mining vendor.
This is partially for flavor reasons, but mostly because whiskey is also
in mining vendors, so miners can make their own monster tamer; they
should be able to do it with the new version, too.
3. Corrected the 20% ABV monster tamer not to be called "vaguely
alcoholic".

## Why It's Good For The Game

1. An alcohol intolerant xenochimera would be a "soft grief" of a sort,
like taking radiation weakness on an engineer character or whatever.
It's possible to play around, but mostly it just causes grief for _other
people_, if it goes wrong.

2. Monster tamer's an _extremely_ goofy part of the setting. It'd be
fine if any ol' drink worked, or if it had to be pure ethanol, but no,
it has to be _specifically_ whiskey. This handwaves it by making it some
complex long-chain carbohydrate thing.

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
add: Dry monster tamer, made with root beer and animal protein
tweak: Mining vendor gets root beer/sarsaparilla/hard root beer
spellcheck: 20% alcohol-by-volume monster tamer is no longer described
as "vaguely alcoholic"
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->