Tweak some alcohol numbers to be less egregious (#6441)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

> it's fullmerged
> next day i first hear complaints

Anyway. This makes it so that:
1. Particularly onerous effects are less likely to happen. They're not
on a cooldown and don't show a message or anything.
2. Having alcohol effects applied to you makes ethanol metabolize
faster.
3. It now takes 42u instead of 36u of ethanol to start taking liver
damage (since a bottle of vodka gives you a total of 40u...)
4. The magical ultra-proof alien drinks are now 300 proof instead of 400
5. Being hydrated makes it metabolize much faster

## Why It's Good For The Game

The numbers weren't final! It needed more testing! Aa!

## Changelog

:cl:
tweak: Alcohol numbers still being tweaked
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->