Library Curtain, Backwall Light, and Asteroid/Cult Event Volume (#6501)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Fixed a curtain in the Atlas library that was closed but the sprite
showed it open, added a light right next to it to illuminate the back
wall, and cut the volume on the asteroid and cult boarding alarms.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
The curtain goes without saying, it's not meant to be that way, I just
put it the right way on. The light was a request on the mapping issue
triage and I just got to it myself, it does help brighten that corner of
the library when the curtains are closed. And for the volume, it was
tremendously loud, if you're just sitting doing something, and not
hearing airlocks opening then suddenly hear the PA ring and get your
ears BLASTED with these sirens, it's not good.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Changelog
Lowered announcement alarms for the meteors and cult events.
Added a light fixture and fixed a curtain on Atlas.
<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
add: light fixture in the back of Atlas library
fix: fixed the Atlas library's curtain
config: changed meteors and cult's event PA announcement alarm volume
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->