Replaces airlocks access vars with access helpers on Triumph (2/3) (#6497)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

This PR is a continuation of #6462 and aims to replace each and every
access var with access helpers. This is the 2nd of 3.

Also adds `research_lab` access helper and grants Pathfinders access to
Rift's RnD (and every map's RnD from now on).

- [x] [Deck 1](https://www.upload.ee/image/16715103/deck_1.png)
- [x] [Deck 2](https://www.upload.ee/image/16715139/deck_2.png)
- [x] [Deck 3](https://www.upload.ee/image/16715188/deck_3.png)
- [x] [Deck 4](https://www.upload.ee/image/16718583/deck_4.png)
- [x] [Flagship](https://www.upload.ee/image/16718647/flagship.png)
- [x] Misc
- [x] Burn engine
- [x] Fission engine
- [x] Rust engine
- [x] Supermatter engine

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

Access vars are obsolete and really complicated for mappers to use.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
add: Added access helpers to every airlock on Triumph.
add: Added additional maintenance airlocks & hatches to Triumph's engine
chamber.
add: Pathfinders now have access to Rift's RnD
del: Removed Triumph's airlocks access vars, and replaced them with
access helpers.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->