Makes encryption heater "efficiency" actually reduce efficiency (and power use) (#6515)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

As of right now, lower efficiency on the encryption heater just stores
the points in the heater until the efficiency goes up. For example,
having 50% efficiency will make the heater have to go up to 200%
progress-to-next-point to make a point, at which point it'll be at...
100%. What this means in practice is that efficiency does _nothing at
all_ and you can completely ignore the system as long as you're willing
to repair the heater repeatedly, up to infinity.

What this does is make it so that efficiency does two things:

1. Actual power used is multiplied by efficiency. This means that if you
have it set to 1,000,000 joules and it's at 50% efficiency, it'll
actually use only 500000 joules. This is actually to _mitigate_...
2. Actual power conversion efficiency is multiplied by... efficiency. In
other words: efficiency is actually efficiency, now. 50% efficiency
means that you're trying to use 1,000,000 joules but you're actually
only using 500,000 and you get points as if you were using 250,000.

## Why It's Good For The Game

No more setting the heater to infinity and then going "haha look at all
these points building up just waiting for me to cool this air down or
literally just drag it to a colder spot"

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
balance: Encryption heater can only be set to use 1e30 watts of power
balance: Encryption heater efficiency now throttles power use
balance: Encryption heater efficiency now actually reduces points gained
per joule
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

Co-authored-by: LordME <58342752+TheLordME@users.noreply.github.com>