explo equipment stuff no 1: emp bow / hardlight bow (#6523)
## About The Pull Request

starts adding the emp bow

## Why It's Good For The Game

as went over in discord; unstandardizing explo's equipment/making it
more utility based reinforces a 'you are not soldiers' type of deal

~~also, ultra wip~~ **good to go**

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
add: emp bow - rechargable ion gun for explo. 32 cranks for 1 shot lol
lmao
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->