Facility Director's/Officer's Sabre changes (#6558)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

- The Officer's Sabre is now a material item, which means it can be
upgraded with better materials. It starts off as plasteel, with a base
damage of around 25 due to funny material calculations. That does mean
the starting damage does get nerfed a little from the original 30. But
it can go up to around 44 if upgraded with durasteel.
- The sabre now has a 60% chance of parrying melee attacks, and a 50%
chance of deflecting projectiles. For reference, the energy sword has a
65% chance of projectile blocking. Should the captain encounter someone
trying to kill them, they can perhaps depend on the sabre to defend
themselves.
- Fixed the sabre not being sharp (now causes bleeding, now can butcher
mobs). Fixed the sabre's sheath not updating their icon.
- Drawing the sabre from a storage (usually the sheathe) plays the
appropriate noise. En-garde!

## Why It's Good For The Game

The sabre isn't really that great of a weapon to use if the facility
director for some reason has to get into a fight. Sure, the FD shouldn't
actively be putting themselves in danger, but it should make things more
interesting should they find themselves in such danger to begin with.
Like getting shot at by Kynde representatives in a meeting room, for
instance.

Central gave the FD a cool sword for some reason. Might as well make it
actually cool.

This probably could also incentivize exploration to try out the colonial
equipment pack some more too, since the sabre does appear in that
equipment pack.

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
balance: The Officer's Sabre is now a material weapon, which means it
can be upgraded. It starts off as plasteel with around 25 damage.
balance: The Officer's Sabre now has a 60% chance to deflect melee
attacks, and 50% chance to deflect projectiles, for as long as you are
facing your enemy.
fix: The sabre sheath updates its icon when removing or adding the
sabre.
fix: The officer's sabre is now sharp like it should be.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->