Adds Donut & Condiment Supply Packs for Cargo (#6566)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

The donut supply crate contains 3 donut boxes.
The condiment supply crate contains several condiments. View the code to
see which ones, since there's a good bit of them. Salt, pepper, and
space spice are provided 3 times.

## Why It's Good For The Game
A few of the condiments included in the condiment box currently isn't
renewable at all during a shift. This is one solution that's provided
until hydroponics gets some plants coded in that can involve that.

As for the donut boxes - donuts are a common thing from stores. It'd
also make sense that NT would let you order more donuts since they tend
to stock the crew with several donuts shiftstart already.


<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
add: Donuts and condiments can now be ordered from Cargo in the
hospitality category.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->