Adds spray bottle & more glasses to autolathe (#6567)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

- Adds spray bottles to the autolathe. This also means spray bottle scan
get recycled.
- since various items are being an extension of spray bottle, they also
get a base material that could be used for recycling or whatever else.
- The metamaterial glass, coffee cup, and fitness cup have also been
added to the autolathe.
- The costs of making these glasses have been made cheaper to make it
consistent with the other glasses available in the lathe.

## Why It's Good For The Game

"Hold up, giving people spray bottles seems like a bad idea."
So is a chem sprayer, and you can make those in R&D.

At minimum, people can fill it with water and make water messes. They
can also do that with a fire extinguisher. In order to cause more chaos,
they'd need to have access to touch chemicals, which is only likely if
you're a security officer, or a job that's in the medical or research
department. Which have access to spraying equipment anyways.

Anyways, you can just make another spray bottle if someone stole all the
spray bottles.

Having the capability to make more metaglasses is good otherwise.


## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
add: The spray bottle and a few extra glasses have been added to the
autolathe.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->