Lunchbox additions, fixes liquid ration problem (#6572)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

- Fixes things not being properly pathed and resulting in an invisible
"snack" where LiquidFood was intended instead.

Adds the following to the lunchbox loadout options:
- LiquidProtein Ration
- Blood Sausage
- Moghesian Sea Delight
- Bludbox and Bludbox Light
- Milk and Chocolate Milk
- Dumb Bitch Juice

## Why It's Good For The Game

This is primarily a PR focused on giving vetalan, xenochimera, unathi,
or just someone who likes blood some lunch options for their lunchbox.
Otherwise, fixes good.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
add: Adds a few new options for the lunchbox loadout item, particularly
blood-related foods for vetalan and xenochimera folk.
fix: A few areas where liquidfood was pathed improperly is now pathed
properly.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->