Makes some things attackable via melee, code tweaks (#6576)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

- Rocks are attackable.
- Outcrops are attackable.
- The two above can also be instantly destroyed by a proto-kinetic
accelerator.
- "Ausbushes" are attackable.
- Floodlights are attackable.
- Curtains are attackable.
- Fences are attackable.
- some other random things are attackable.
- gives feedback when using the fossil machine
- removed redundant code from curtains

## Why It's Good For The Game

get this snowflake no-melee nonsense out of my way.

there's probably a lot more to be sorted out but some of these are the
more egregious examples because they're usually obstructive in some way.

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
fix: A few rocks, plants, and some other things are able to be removed
via harm intent melee. You can now blast rocks with PKAs as well.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->