Refactor rotation code in DMM loader
General refactor of how we handle rotation in DMM loading.

Previously:

- A lot of unnecessary code to thread and repeatedly add
  orientation-related variables to the preloader.
- dmm_parsed previously modified its own args, then used arglist(args)
  This is bad energy -- as far as I can tell, this was was relying on
  the fact that _load_impl didn't use the modified values.
- Fire alarms were hardcoded to deal with east/west incorrectly

Now:

- These variables now live in the dmm_context.
- _load_impl takes fewer unnecessary args, and all its args are explicit
- A few variables have been removed or had their scope limited.
- Fire alarms aren't hardcoded to do that anymore.

I haven't done any profiling, but I'll update this with informal metrics
after I know that the code is correct. Back of the envelope calculation
implies the number of variable-on-object lookups in the loop should be pretty
close to the original number.

The one lookup that looks new (where we get the context from the preloader
in preloader_load) actually happened unconditionally at the end of the
procedure in the previous code.