Swaps Riftmap's AI Uplink areas to the correct configuration and replaces the engineering sign in atmospherics to atmospherics (#6554)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Swaps Riftmap's AI Uplink areas to the correct configuration and
replaces the engineering sign in atmospherics to atmospherics, and
bugfixes, hide the pipes, hide the bridge
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Atmospherics is well tagged, so people stop thinking it's engineering.
And the areas are fixed so they match their room.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Changelog
Swaps Atlas' Cyborg Station Area with Messaging Server Area between
them. Changes the sign in front of atmospherics to the correct one.
Added a name to the entrance airlock into atmospherics. Also some bug
fixes.
<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
add: added the name of "Atmospherics Airlock" to the entrance airlock
into atmospherics
fix: replaced the engineering sign in front of atmospherics with the
atmospherics one
fix: swapped ai borg station and ai server so that they're in the right
spots
fix: hid four atmos pipes at the entrance into maints from library
corridor in riftmap
fix: made a window in riftmap's bridge electrochromatic
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->