Fix hovering fire alarms, refactor DMM rotation code (#6589)
## About The Pull Request

Previously:

- A lot of unnecessary code to thread and repeatedly add
orientation-related variables to the preloader.
- dmm_parsed previously modified its own args, then used arglist(args)
This is bad energy -- as far as I can tell, this was was relying on the
fact that _load_impl didn't use the modified values.
- Fire alarms were hardcoded to deal with east/west incorrectly

Now:

- These variables now live in the dmm_context.
- _load_impl takes fewer unnecessary args, and all its args are explicit
- A few variables have been removed or had their scope limited.
- Fire alarms aren't hardcoded to do that anymore.

Bugs I didn't fix:

- When a map is rotated EAST or WEST, signs spawn in an overlapping
fashion that's really ugly.

## Why It's Good For The Game

### Correctness

Fire alarms were spawning sideways when rotated EAST or WEST and
probably shouldn't be. I've fixed this behavior now.

### Style

Threading a lot of variables in DM is pretty rough -- DM doesn't
actually warn if you do this wrong. Silicons specifically flagged this
as a rough spot for code quality, which is why I attacked it.

These variables never change during the load process, so threading them
as arguments creates the false implication that they can be different
per map object.

### Performance

My goal wasn't to make performance better, just to make the code more
readable without making it worse. Brief profiling implies it might be
faster, but the difference is marginal and possibly explained by bad
test conditions:

- map loading took 24.0, 24.3, 24.6 seconds before the change
- map loading took 23.0, 23.6, 23.8 seconds after this change

All tests were done with rotated Triumph. (I haven't profiled
non-rotated yet.)

Back of the envelope calculation implies the number of
variable-on-object lookups in the hot loop phase of the map load should
be pretty close to the original number -- maybe a little lower, since
I've turned some into locals.

The one lookup that looks new (where we get the context from the
preloader in preloader_load) was actually one that happened
unconditionally at the end of the procedure in the previous code.

## Images

Sign bug: I didn't fix this or introduce this.


![image](https://github.com/user-attachments/assets/47debb5e-00e4-4bf7-a404-94d315f07607)

Fire alarm bug living its best life. I fixed this:


![image](https://github.com/user-attachments/assets/f1db22d4-921c-4a1d-827f-cb49c76a2cb9)

Fire alarm is Tethered. (Post-fix):


![image](https://github.com/user-attachments/assets/b939842f-8dfc-45c3-bfcb-e6faf3a90ee8)

## Changelog

:cl: Pyrex
refactor: DMM rotation code is a little tidier
fix: fire alarms have a 0% chance of hovering
/:cl:

---------

Co-authored-by: GDN <96800819+GDNgit@users.noreply.github.com>