Pre-burns the mineral planet (#6590)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

A planet can't have an atmosphere filled with fuel and oxidizer at the
same time, except trace amounts of one or the other. If a planet's
atmosphere is sufficient for fire to spread, then a single spark of any
sort anywhere (such as, say, lightning, or a rock rolling down a hill
just right) will cause the whole planet to catch fire.

This changes the Class G mineral planet so that it's already done that.

## Why It's Good For The Game

I mean having a burn mix planet is genuinely a little funny but it's:

1. Not actually that fun gameplay-wise (it just means nobody ever goes
there)
2. Completely goofy setting-wise (for reasons mentioned)
3. Phoron's awful on its own, it doesn't need to catch fire

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
tweak: Oxygen on the Class G Mineral-Rich planet changed to CO2
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->