non-tgui age gate (#6546)
## About The Pull Request
title

our options are:
1. wait an arbitrary amount of time and then make the tgui input
2. use a non-tgui age gate

testmerge only for the sake of i want to be able to throw a test account
at it on the server which has DB access because i'm not setting up a
local db connection for this (see: i'm lazy)

## Why It's Good For The Game
fix issues with the prompt loading with tgui

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
fix: fixes loading issues with the age gate (new players can just get
stuck and unable to join)
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->