New features, New wrappers for cell_slot system. (#6421)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Adds wrappers for cells to the cell_slot system for cleaner calling.
Adds a function to remove a cell and hand it directly to a mob for saner
handling outside the file. Adds a bunch of comments from what I've
divined in the hopes of granting others the same wisdom.
Swaps the single inducer line that can use the new wrapper procs.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Will hopefully allow others to look upon silicons work and understand it
easier. Allows me to write prettier and safer code for the vehicle
refactor.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Changelog
-Comments a bunch of the functions because my brain is smooth
-Adds wrappers for cell.dm functions to allow for cleaner code lines
when dealing with this system
-Adds mob_remove_cell() for easier cell removal in vehicles update.
-"Updates" inducers to use the new wrappers (a single proc)
<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl: Shadowcat
code: Adds cell.dm function wrappers to cell_slot.dm
code: Adds a function to remove a cell and hand it directly to a move in
cell_slot.dm
refactor: Comments a lot of the functions in cell_slot.dm
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Co-authored-by: LordME <58342752+TheLordME@users.noreply.github.com>